@@ -1,4 +1,6 @@
-﻿using System.Net.Http;
+﻿using System.Diagnostics.CodeAnalysis;
+using System.IO;
+using System.Net.Http;
 using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
@@ -7,17 +9,21 @@ using Newtonsoft.Json;
 
 namespace Wally.RomMaster.FileService.Tests.IntegrationTests.Helpers;
 
+[SuppressMessage("Major Code Smell", "S4005:\"System.Uri\" arguments should be used instead of strings")]
 public static class HttpClientExtensions
 {
 	private const string _jsonMediaType = "application/json";
 
 	private static readonly JsonSerializerSettings _jsonSettings =
-		new() { ContractResolver = new PrivateSetterContractResolver(), };
+		new()
+		{
+			ContractResolver = new PrivateSetterContractResolver(),
+		};
 
-	public static Task<HttpResponseMessage> PutAsync(
+	public static Task<HttpResponseMessage> PutAsync<TPayload>(
 		this HttpClient client,
 		string url,
-		object payload,
+		TPayload payload,
 		CancellationToken cancellationToken)
 	{
 		var content = CreateContent(payload);
@@ -25,10 +31,10 @@ public static class HttpClientExtensions
 		return client.PutAsync(url, content, cancellationToken);
 	}
 
-	public static Task<HttpResponseMessage> PostAsync(
+	public static Task<HttpResponseMessage> PostAsync<TPayload>(
 		this HttpClient client,
 		string url,
-		object payload,
+		TPayload payload,
 		CancellationToken cancellationToken)
 	{
 		var content = CreateContent(payload);
@@ -36,15 +42,20 @@ public static class HttpClientExtensions
 		return client.PostAsync(url, content, cancellationToken);
 	}
 
-	public static async Task<T> ReadAsync<T>(this HttpResponseMessage response, CancellationToken cancellationToken)
+	public static async Task<TResponse> ReadAsync<TResponse>(this HttpResponseMessage response,
+		CancellationToken cancellationToken)
 	{
-		var json = await response.Content.ReadAsStringAsync(
-			cancellationToken); // TODO: get Stream instead of loading String
-		return JsonConvert.DeserializeObject<T>(json, _jsonSettings) !;
+		var stream = await response.Content.ReadAsStreamAsync(cancellationToken);
+
+		using TextReader textReader = new StreamReader(stream);
+		using JsonReader jsonReader = new JsonTextReader(textReader);
+
+		return JsonSerializer.Create(_jsonSettings)
+			.Deserialize<TResponse>(jsonReader) !;
 	}
 
-	private static StringContent CreateContent<T>(T item)
+	private static StringContent CreateContent<TPayload>(TPayload payload)
 	{
-		return new StringContent(JsonConvert.SerializeObject(item), Encoding.UTF8, _jsonMediaType);
+		return new StringContent(JsonConvert.SerializeObject(payload), Encoding.UTF8, _jsonMediaType);
 	}
 }
