@@ -1,7 +1,7 @@
 ﻿using System.Threading;
 using System.Threading.Tasks;
-using Wally.Lib.DDD.Abstractions.Queries;
 using Wally.RomMaster.FileService.Application.Contracts.Responses.Users;
+using Wally.Lib.DDD.Abstractions.Queries;
 
 namespace Wally.RomMaster.FileService.Application.Users.Queries;
 
