@@ -2,12 +2,10 @@
 
 namespace Wally.RomMaster.FileService.Domain.Abstractions;
 
-public abstract class GuidId<TId> : StronglyTypedId<TId, Guid>
+public class GuidId<TId> : StronglyTypedId<TId, Guid>
 	where TId : StronglyTypedId<TId, Guid>
 {
-	private static readonly long _baseDateTicks = new DateTime(2000, 1, 1).Ticks;
-
-	public GuidId()
+	protected GuidId()
 		: this(Generate())
 	{
 	}
@@ -16,11 +14,6 @@ public abstract class GuidId<TId> : StronglyTypedId<TId, Guid>
 		: base(value)
 	{
 	}
-
-	/*public static explicit operator Guid(UserId id)
-	{
-		return id.Value;
-	}*/
 
 	/// <summary>
 	///     https://github.com/nhibernate/nhibernate-core/blob/master/src/NHibernate/Id/GuidCombGenerator.cs
@@ -34,20 +27,20 @@ public abstract class GuidId<TId> : StronglyTypedId<TId, Guid>
 
 		var now = DateTime.UtcNow;
 
-		// Get the days and milliseconds which will be used to build the byte string 
-		var days = new TimeSpan(now.Ticks - _baseDateTicks);
+		// Get the days and milliseconds which will be used to build the byte string
+		var days = new TimeSpan(now.Ticks - DateTime.UnixEpoch.Ticks);
 		var msecs = now.TimeOfDay;
 
-		// Convert to a byte array 
-		// Note that SQL Server is accurate to 1/300th of a millisecond so we divide by 3.333333 
+		// Convert to a byte array
+		// Note that SQL Server is accurate to 1/300th of a millisecond so we divide by 3.333333
 		var daysArray = BitConverter.GetBytes(days.Days);
 		var msecsArray = BitConverter.GetBytes((long)(msecs.TotalMilliseconds / 3.333333));
 
-		// Reverse the bytes to match SQL Servers ordering 
+		// Reverse the bytes to match SQL Servers ordering
 		Array.Reverse(daysArray);
 		Array.Reverse(msecsArray);
 
-		// Copy the bytes into the guid 
+		// Copy the bytes into the guid
 		Array.Copy(daysArray, daysArray.Length - 2, guidArray, guidArray.Length - 6, 2);
 		Array.Copy(msecsArray, msecsArray.Length - 4, guidArray, guidArray.Length - 4, 4);
 
