@@ -5,10 +5,10 @@ using System.Linq;
 
 namespace Wally.RomMaster.FileService.Domain.Abstractions;
 
-// TODO: https://github.com/dotnet/efcore/blob/release/8.0/src/EFCore/ValueGeneration/SequentialGuidValueGenerator.cs
-
+// https://github.com/dotnet/efcore/blob/release/8.0/src/EFCore/ValueGeneration/SequentialGuidValueGenerator.cs
+// https://andrewlock.net/series/using-strongly-typed-entity-ids-to-avoid-primitive-obsession/
 [TypeConverter(typeof(StronglyTypedIdConverter))]
-public abstract class StronglyTypedId<TStronglyTypedId, TValue> : IStronglyTypedId<TStronglyTypedId, TValue>
+public class StronglyTypedId<TStronglyTypedId, TValue> : IStronglyTypedId<TStronglyTypedId, TValue>
 	where TStronglyTypedId : StronglyTypedId<TStronglyTypedId, TValue>
 	where TValue : notnull, IComparable
 {
@@ -19,20 +19,12 @@ public abstract class StronglyTypedId<TStronglyTypedId, TValue> : IStronglyTyped
 	/// <remarks>
 	///     See http://computinglife.wordpress.com/2008/11/20/why-do-hash-functions-use-prime-numbers/
 	/// </remarks>
-	private const int HashMultiplier = 37;
-
-	/*static StronglyTypedId()
-	{
-		Type valueType = typeof(TValue);
-		bool isIdentifierType = valueType.IsNumeric() || valueType == typeof(string) || valueType == typeof(Guid);
-
-		// Guard.Against.False(isIdentifierType, nameof(Value), "The value of a strongly-typed ID must be a numeric, string or Guid type.");
-	}*/
+	private const int _hashMultiplier = 37;
 
 	/// <summary>
-	///     Initializes a new instance of the <see cref="StronglyTypedId{TStronglyTypedId,TValue}" /> type.
+	///     Initializes a new instance of the <see cref="StronglyTypedId{TStronglyTypedId, TValue}" /> class.
 	/// </summary>
-	/// <param name="value"></param>
+	/// <param name="value">The Value.</param>
 	protected StronglyTypedId(TValue value)
 	{
 		Value = value;
@@ -40,37 +32,56 @@ public abstract class StronglyTypedId<TStronglyTypedId, TValue> : IStronglyTyped
 
 	public TValue Value { get; }
 
-	public bool Equals(TStronglyTypedId other)
-	{
-		return Equals(other as object);
-	}
-
 	public int CompareTo(TStronglyTypedId? other)
 	{
+		if (other is null)
+		{
+			return 1;
+		}
+
 		return (Value, other.Value) switch
 		{
 			(null, null) => 0,
 			(null, _) => -1,
 			(_, null) => 1,
-			(_, _) => Value.CompareTo(other.Value),
+			_ => Value.CompareTo(other.Value),
 		};
 	}
 
-	public sealed override bool Equals(object? obj)
+	public bool Equals(TStronglyTypedId? other)
 	{
-		if (obj is null)
+		if (other is null)
 		{
 			return false;
 		}
 
-		if (ReferenceEquals(this, obj))
+		if (ReferenceEquals(this, other))
 		{
 			return true;
 		}
 
-		var other = obj as StronglyTypedId<TStronglyTypedId, TValue>;
-		return other != null && GetType() == other.GetType() && GetEqualityComponents()
+		return GetType() == other.GetType() && GetEqualityComponents()
 			.SequenceEqual(other.GetEqualityComponents());
+	}
+
+	public bool Equals(TStronglyTypedId? x, TStronglyTypedId? y)
+	{
+		if (x is null)
+		{
+			return y is null;
+		}
+
+		return x.Equals(y);
+	}
+
+	public int GetHashCode(TStronglyTypedId obj)
+	{
+		return obj.GetHashCode();
+	}
+
+	public sealed override bool Equals(object? obj)
+	{
+		return Equals(obj as TStronglyTypedId);
 	}
 
 	/// <summary>
@@ -97,10 +108,7 @@ public abstract class StronglyTypedId<TStronglyTypedId, TValue> : IStronglyTyped
 
 			foreach (var component in GetEqualityComponents())
 			{
-				if (component != null)
-				{
-					hashCode = (hashCode * HashMultiplier) ^ component.GetHashCode();
-				}
+				hashCode = (hashCode * _hashMultiplier) ^ component.GetHashCode();
 			}
 
 			return hashCode;
@@ -132,44 +140,30 @@ public abstract class StronglyTypedId<TStronglyTypedId, TValue> : IStronglyTyped
 		return !(left == right);
 	}
 
-	/*/// <summary>
-	///     Converts a value implicitly to an instance of TStronglyTypedId.
-	/// </summary>
-	/// <param name="value">The value</param>
-	public static explicit operator StronglyTypedId<TStronglyTypedId, TValue>(TValue value)
-	{
-		object instance = Activator.CreateInstance(typeof(TStronglyTypedId), new object[] { value });
-		return (TStronglyTypedId)instance;
-	}
-
-	public static implicit operator TValue(StronglyTypedId<TStronglyTypedId, TValue> id)
-	{
-		return id.Value;
-	}*/
-
 	/// <summary>
 	///     Converts a value implicitly to an instance of TStronglyTypedId.
 	/// </summary>
-	/// <param name="value">The value</param>
+	/// <param name="value">The value.</param>
 	public static explicit operator StronglyTypedId<TStronglyTypedId, TValue>(TValue value)
 	{
 		var instance = Activator.CreateInstance(typeof(TStronglyTypedId), value) !;
 		return (TStronglyTypedId)instance;
 	}
 
-	/*public static explicit operator TValue(StronglyTypedId<TStronglyTypedId, TValue> id)
-	{
-		return id.Value;
-	}*/
-
-	public override string ToString()
+	public override string? ToString()
 	{
 		return Value.ToString();
 	}
 }
 
 public interface IStronglyTypedId
-	<TStronglyTypedId, out TKey> : IComparable<TStronglyTypedId>, IEquatable<TStronglyTypedId>
+	<TStronglyTypedId, out TKey> : IStronglyTypedId<TKey>, IComparable<TStronglyTypedId>, IEquatable<TStronglyTypedId>,
+		IEqualityComparer<TStronglyTypedId>
+	where TKey : notnull, IComparable
+{
+}
+
+public interface IStronglyTypedId<out TKey>
 	where TKey : notnull, IComparable
 {
 	/// <summary>
@@ -180,6 +174,43 @@ public interface IStronglyTypedId
 
 public static class TypeExtensions
 {
+	private static readonly Dictionary<Type, bool> _numericTypes = new()
+	{
+		{
+			typeof(sbyte), true
+		},
+		{
+			typeof(byte), true
+		},
+		{
+			typeof(short), true
+		},
+		{
+			typeof(ushort), true
+		},
+		{
+			typeof(int), true
+		},
+		{
+			typeof(uint), true
+		},
+		{
+			typeof(long), true
+		},
+		{
+			typeof(ulong), true
+		},
+		{
+			typeof(decimal), true
+		},
+		{
+			typeof(float), true
+		},
+		{
+			typeof(double), true
+		},
+	};
+
 	/// <summary>
 	///     Determines if the given type is numeric.
 	/// </summary>
@@ -190,9 +221,8 @@ public static class TypeExtensions
 	public static bool IsNumeric(this Type type)
 	{
 		type = type.UnwrapNullableType();
-		return type == typeof(sbyte) || type == typeof(byte) || type == typeof(short) || type == typeof(ushort) ||
-			type == typeof(int) || type == typeof(uint) || type == typeof(long) || type == typeof(ulong) ||
-			type == typeof(decimal) || type == typeof(float) || type == typeof(double);
+
+		return _numericTypes.ContainsKey(type) && _numericTypes[type];
 	}
 
 	/// <summary>
