@@ -1,15 +1,14 @@
 ﻿using System;
+using System.Diagnostics.CodeAnalysis;
 using System.Threading;
 using System.Threading.Tasks;
-using Confluent.Kafka;
 using MassTransit;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
-using Wally.RomMaster.FileService.Application.Messages.Files;
+using Wally.RomMaster.FileService.Application.Messages.Users;
 using Wally.RomMaster.FileService.Infrastructure.DI.Microsoft.Models;
 using Wally.RomMaster.FileService.Infrastructure.Messaging;
 using Wally.RomMaster.FileService.Infrastructure.Messaging.Consumers;
-using Wally.RomMaster.HashService.Application.Messages.Hashes;
 
 namespace Wally.RomMaster.FileService.Infrastructure.DI.Microsoft.Extensions;
 
@@ -37,41 +36,31 @@ public static class MessagingExtensions
 						break;
 					case MessageBrokerType.Kafka:
 						a.UsingInMemory((context, config) => config.ConfigureEndpoints(context));
+						a.AddRider(
+							rider =>
+							{
+								rider.AddConsumersFromNamespaceContaining<IInfrastructureMessagingAssemblyMarker>();
 
-						// a.UsingRabbitMq((context, cfg) => cfg.ConfigureEndpoints(context));
+								// TODO: auto-register
+								rider.AddProducer<UserCreatedMessage>(nameof(UserCreatedMessage));
 
-						a.AddRider(rider =>
-						{
-							rider.AddConsumersFromNamespaceContaining<IInfrastructureMessagingAssemblyMarker>();
-							// TODO: auto-register
-							// rider.AddConsumer<KafkaMessageConsumer>();
-							// rider.AddProducer<FileCreatedMessage>(nameof(FileCreatedMessage));
-							// rider.AddProducer<FileModifiedMessage>(nameof(FileModifiedMessage));
-							rider.AddProducer<FileCreatedMessage>(typeof(FileCreatedMessage).FullName);
-							rider.AddProducer<FileModifiedMessage>(typeof(FileModifiedMessage).FullName);
+								// rider.AddProducer<UserUpdatedMessage>(nameof(FileModifiedMessage));
 
-							rider.UsingKafka(
-								(context, k) =>
-								{
-									k.ClientId = typeof(IInfrastructureMessagingAssemblyMarker).Namespace;
-									k.Host(settings.ConnectionStrings.ServiceBus);
-
-									// TODO: auto-register
-									/*k.TopicEndpoint<KafkaMessage>(typeof(IInfrastructureMessagingAssemblyMarker).Namespace, typeof(IInfrastructureMessagingAssemblyMarker).Namespace, e =>
+								rider.UsingKafka(
+									(context, k) =>
 									{
-										e.ConfigureConsumer<KafkaMessageConsumer>(context);
-									});*/
-									k.TopicEndpoint<HashComputedMessage>(typeof(HashComputedMessage).FullName,
-										typeof(IInfrastructureMessagingAssemblyMarker).Namespace,
-										e =>
-										{
-											e.AutoOffsetReset = AutoOffsetReset.Earliest;
-											e.ConfigureConsumer<HashComputedMessageConsumer>(context);
-										});
-								});
+										k.ClientId = typeof(IInfrastructureMessagingAssemblyMarker).Namespace;
+										k.Host(settings.ConnectionStrings.ServiceBus);
 
-							services.AddScoped<IBus, KafkaBus>();
-						});
+										// TODO: auto-register
+										k.TopicEndpoint<UserCreatedMessage>(
+											typeof(IInfrastructureMessagingAssemblyMarker).Namespace,
+											typeof(IInfrastructureMessagingAssemblyMarker).Namespace,
+											e => { e.ConfigureConsumer<UserCreatedMessageConsumer>(context); });
+									});
+
+								services.AddScoped<IBus, KafkaBus>();
+							});
 						break;
 					case MessageBrokerType.RabbitMQ:
 						a.UsingRabbitMq(
@@ -82,8 +71,7 @@ public static class MessagingExtensions
 							});
 						break;
 					default:
-						throw new ArgumentOutOfRangeException(
-							nameof(settings.MessageBroker),
+						throw new ArgumentOutOfRangeException(nameof(settings.MessageBroker),
 							$"Unknown Message Broker: '{settings.MessageBroker}'");
 				}
 			});
@@ -91,7 +79,8 @@ public static class MessagingExtensions
 		return services;
 	}
 
-	private class BusStub : IBus
+	[SuppressMessage("Major Code Smell", "S4017:Method signatures should not contain nested generic types")]
+	private sealed class BusStub : IBus
 	{
 		private readonly ILogger<BusStub> _logger;
 
@@ -102,13 +91,13 @@ public static class MessagingExtensions
 
 		public ConnectHandle ConnectPublishObserver(IPublishObserver observer)
 		{
-			throw new NotImplementedException();
+			throw new NotSupportedException();
 		}
 
 		public Task<ISendEndpoint> GetPublishSendEndpoint<T>()
 			where T : class
 		{
-			throw new NotImplementedException();
+			throw new NotSupportedException();
 		}
 
 		public Task Publish<T>(T message, CancellationToken cancellationToken = new())
@@ -125,7 +114,7 @@ public static class MessagingExtensions
 			CancellationToken cancellationToken = new())
 			where T : class
 		{
-			throw new NotImplementedException();
+			throw new NotSupportedException();
 		}
 
 		public Task Publish<T>(
@@ -134,12 +123,12 @@ public static class MessagingExtensions
 			CancellationToken cancellationToken = new())
 			where T : class
 		{
-			throw new NotImplementedException();
+			throw new NotSupportedException();
 		}
 
 		public Task Publish(object message, CancellationToken cancellationToken = new())
 		{
-			throw new NotImplementedException();
+			throw new NotSupportedException();
 		}
 
 		public Task Publish(
@@ -147,12 +136,12 @@ public static class MessagingExtensions
 			IPipe<PublishContext> publishPipe,
 			CancellationToken cancellationToken = new())
 		{
-			throw new NotImplementedException();
+			throw new NotSupportedException();
 		}
 
 		public Task Publish(object message, Type messageType, CancellationToken cancellationToken = new())
 		{
-			throw new NotImplementedException();
+			throw new NotSupportedException();
 		}
 
 		public Task Publish(
@@ -161,13 +150,13 @@ public static class MessagingExtensions
 			IPipe<PublishContext> publishPipe,
 			CancellationToken cancellationToken = new())
 		{
-			throw new NotImplementedException();
+			throw new NotSupportedException();
 		}
 
 		public Task Publish<T>(object values, CancellationToken cancellationToken = new())
 			where T : class
 		{
-			throw new NotImplementedException();
+			throw new NotSupportedException();
 		}
 
 		public Task Publish<T>(
@@ -176,7 +165,7 @@ public static class MessagingExtensions
 			CancellationToken cancellationToken = new())
 			where T : class
 		{
-			throw new NotImplementedException();
+			throw new NotSupportedException();
 		}
 
 		public Task Publish<T>(
@@ -185,61 +174,61 @@ public static class MessagingExtensions
 			CancellationToken cancellationToken = new())
 			where T : class
 		{
-			throw new NotImplementedException();
+			throw new NotSupportedException();
 		}
 
 		public ConnectHandle ConnectSendObserver(ISendObserver observer)
 		{
-			throw new NotImplementedException();
+			throw new NotSupportedException();
 		}
 
 		public Task<ISendEndpoint> GetSendEndpoint(Uri address)
 		{
-			throw new NotImplementedException();
+			throw new NotSupportedException();
 		}
 
 		public ConnectHandle ConnectConsumePipe<T>(IPipe<ConsumeContext<T>> pipe)
 			where T : class
 		{
-			throw new NotImplementedException();
+			throw new NotSupportedException();
 		}
 
 		public ConnectHandle ConnectConsumePipe<T>(IPipe<ConsumeContext<T>> pipe, ConnectPipeOptions options)
 			where T : class
 		{
-			throw new NotImplementedException();
+			throw new NotSupportedException();
 		}
 
 		public ConnectHandle ConnectRequestPipe<T>(Guid requestId, IPipe<ConsumeContext<T>> pipe)
 			where T : class
 		{
-			throw new NotImplementedException();
+			throw new NotSupportedException();
 		}
 
 		public ConnectHandle ConnectConsumeMessageObserver<T>(IConsumeMessageObserver<T> observer)
 			where T : class
 		{
-			throw new NotImplementedException();
+			throw new NotSupportedException();
 		}
 
 		public ConnectHandle ConnectConsumeObserver(IConsumeObserver observer)
 		{
-			throw new NotImplementedException();
+			throw new NotSupportedException();
 		}
 
 		public ConnectHandle ConnectReceiveObserver(IReceiveObserver observer)
 		{
-			throw new NotImplementedException();
+			throw new NotSupportedException();
 		}
 
 		public ConnectHandle ConnectReceiveEndpointObserver(IReceiveEndpointObserver observer)
 		{
-			throw new NotImplementedException();
+			throw new NotSupportedException();
 		}
 
 		public ConnectHandle ConnectEndpointConfigurationObserver(IEndpointConfigurationObserver observer)
 		{
-			throw new NotImplementedException();
+			throw new NotSupportedException();
 		}
 
 		public HostReceiveEndpointHandle ConnectReceiveEndpoint(
@@ -247,27 +236,28 @@ public static class MessagingExtensions
 			IEndpointNameFormatter? endpointNameFormatter = null,
 			Action<IReceiveEndpointConfigurator>? configureEndpoint = null)
 		{
-			throw new NotImplementedException();
+			throw new NotSupportedException();
 		}
 
 		public HostReceiveEndpointHandle ConnectReceiveEndpoint(
 			string queueName,
 			Action<IReceiveEndpointConfigurator>? configureEndpoint = null)
 		{
-			throw new NotImplementedException();
+			throw new NotSupportedException();
 		}
 
 		public void Probe(ProbeContext context)
 		{
-			throw new NotImplementedException();
+			throw new NotSupportedException();
 		}
 
-		public Uri Address { get; }
+		public Uri Address { get; } = null!;
 
-		public IBusTopology Topology { get; }
+		public IBusTopology Topology { get; } = null!;
 	}
 
-	private class KafkaBus : IBus
+	[SuppressMessage("Major Code Smell", "S4017:Method signatures should not contain nested generic types")]
+	private sealed class KafkaBus : IBus
 	{
 		private readonly ILogger<KafkaBus> _logger;
 		private readonly IServiceProvider _serviceProvider;
@@ -280,22 +270,21 @@ public static class MessagingExtensions
 
 		public ConnectHandle ConnectPublishObserver(IPublishObserver observer)
 		{
-			throw new NotImplementedException();
+			throw new NotSupportedException();
 		}
 
 		public Task<ISendEndpoint> GetPublishSendEndpoint<T>()
 			where T : class
 		{
-			throw new NotImplementedException();
+			throw new NotSupportedException();
 		}
 
 		public async Task Publish<T>(T message, CancellationToken cancellationToken = new())
 			where T : class
 		{
+			_logger.LogDebug("Publishing {FullName}", typeof(T).FullName);
+
 			var topicProducer = _serviceProvider.GetRequiredService<ITopicProducer<T>>();
-			
-			_logger.LogInformation($"Produce: '{message.GetType().Name}'.");
-			
 			await topicProducer.Produce(message, cancellationToken);
 		}
 
@@ -305,7 +294,7 @@ public static class MessagingExtensions
 			CancellationToken cancellationToken = new())
 			where T : class
 		{
-			throw new NotImplementedException();
+			throw new NotSupportedException();
 		}
 
 		public Task Publish<T>(
@@ -314,12 +303,12 @@ public static class MessagingExtensions
 			CancellationToken cancellationToken = new())
 			where T : class
 		{
-			throw new NotImplementedException();
+			throw new NotSupportedException();
 		}
 
 		public Task Publish(object message, CancellationToken cancellationToken = new())
 		{
-			throw new NotImplementedException();
+			throw new NotSupportedException();
 		}
 
 		public Task Publish(
@@ -327,12 +316,12 @@ public static class MessagingExtensions
 			IPipe<PublishContext> publishPipe,
 			CancellationToken cancellationToken = new())
 		{
-			throw new NotImplementedException();
+			throw new NotSupportedException();
 		}
 
 		public Task Publish(object message, Type messageType, CancellationToken cancellationToken = new())
 		{
-			throw new NotImplementedException();
+			throw new NotSupportedException();
 		}
 
 		public Task Publish(
@@ -341,13 +330,13 @@ public static class MessagingExtensions
 			IPipe<PublishContext> publishPipe,
 			CancellationToken cancellationToken = new())
 		{
-			throw new NotImplementedException();
+			throw new NotSupportedException();
 		}
 
 		public Task Publish<T>(object values, CancellationToken cancellationToken = new())
 			where T : class
 		{
-			throw new NotImplementedException();
+			throw new NotSupportedException();
 		}
 
 		public Task Publish<T>(
@@ -356,7 +345,7 @@ public static class MessagingExtensions
 			CancellationToken cancellationToken = new())
 			where T : class
 		{
-			throw new NotImplementedException();
+			throw new NotSupportedException();
 		}
 
 		public Task Publish<T>(
@@ -365,61 +354,61 @@ public static class MessagingExtensions
 			CancellationToken cancellationToken = new())
 			where T : class
 		{
-			throw new NotImplementedException();
+			throw new NotSupportedException();
 		}
 
 		public ConnectHandle ConnectSendObserver(ISendObserver observer)
 		{
-			throw new NotImplementedException();
+			throw new NotSupportedException();
 		}
 
 		public Task<ISendEndpoint> GetSendEndpoint(Uri address)
 		{
-			throw new NotImplementedException();
+			throw new NotSupportedException();
 		}
 
 		public ConnectHandle ConnectConsumePipe<T>(IPipe<ConsumeContext<T>> pipe)
 			where T : class
 		{
-			throw new NotImplementedException();
+			throw new NotSupportedException();
 		}
 
 		public ConnectHandle ConnectConsumePipe<T>(IPipe<ConsumeContext<T>> pipe, ConnectPipeOptions options)
 			where T : class
 		{
-			throw new NotImplementedException();
+			throw new NotSupportedException();
 		}
 
 		public ConnectHandle ConnectRequestPipe<T>(Guid requestId, IPipe<ConsumeContext<T>> pipe)
 			where T : class
 		{
-			throw new NotImplementedException();
+			throw new NotSupportedException();
 		}
 
 		public ConnectHandle ConnectConsumeMessageObserver<T>(IConsumeMessageObserver<T> observer)
 			where T : class
 		{
-			throw new NotImplementedException();
+			throw new NotSupportedException();
 		}
 
 		public ConnectHandle ConnectConsumeObserver(IConsumeObserver observer)
 		{
-			throw new NotImplementedException();
+			throw new NotSupportedException();
 		}
 
 		public ConnectHandle ConnectReceiveObserver(IReceiveObserver observer)
 		{
-			throw new NotImplementedException();
+			throw new NotSupportedException();
 		}
 
 		public ConnectHandle ConnectReceiveEndpointObserver(IReceiveEndpointObserver observer)
 		{
-			throw new NotImplementedException();
+			throw new NotSupportedException();
 		}
 
 		public ConnectHandle ConnectEndpointConfigurationObserver(IEndpointConfigurationObserver observer)
 		{
-			throw new NotImplementedException();
+			throw new NotSupportedException();
 		}
 
 		public HostReceiveEndpointHandle ConnectReceiveEndpoint(
@@ -427,23 +416,23 @@ public static class MessagingExtensions
 			IEndpointNameFormatter? endpointNameFormatter = null,
 			Action<IReceiveEndpointConfigurator>? configureEndpoint = null)
 		{
-			throw new NotImplementedException();
+			throw new NotSupportedException();
 		}
 
 		public HostReceiveEndpointHandle ConnectReceiveEndpoint(
 			string queueName,
 			Action<IReceiveEndpointConfigurator>? configureEndpoint = null)
 		{
-			throw new NotImplementedException();
+			throw new NotSupportedException();
 		}
 
 		public void Probe(ProbeContext context)
 		{
-			throw new NotImplementedException();
+			throw new NotSupportedException();
 		}
 
-		public Uri Address { get; }
+		public Uri Address { get; } = null!;
 
-		public IBusTopology Topology { get; }
+		public IBusTopology Topology { get; } = null!;
 	}
 }
