@@ -20,13 +20,12 @@ public static class ServiceCollectionExtensions
 		services.AddOptions(settings);
 		services.AddWebApi();
 		services.AddCqrs();
-		services.AddSwagger(Assembly.GetCallingAssembly());
+		services.AddOpenApi(Assembly.GetCallingAssembly());
 		services.AddHealthChecks(settings);
-		services.AddDbContext(settings);
+		services.AddAddPersistence(settings);
 		services.AddMapper();
 		services.AddMessaging(settings);
 		services.AddEventHub();
-		// services.AddBackgroundServices(settings); // This applying is postponed in Program.cs 
 
 		return services;
 	}
@@ -51,7 +50,7 @@ public static class ServiceCollectionExtensions
 
 		app.UseWebApi();
 
-		app.UseDbContext();
+		app.UsePersistence();
 		app.UseEventHub<EventHub>();
 
 		return app;
