@@ -1,9 +1,9 @@
 using System.Threading.Tasks;
 using MassTransit;
 using MediatR;
-using Wally.Identity.Messages.Users;
 using Wally.RomMaster.FileService.Application.Users.Commands;
 using Wally.RomMaster.FileService.Domain.Users;
+using Wally.Identity.Messages.Users;
 
 namespace Wally.RomMaster.FileService.Infrastructure.Messaging.Consumers;
 
