@@ -1,5 +1,5 @@
 ﻿namespace Wally.RomMaster.FileService.Infrastructure.Persistence.PostgreSQL;
 
-public interface IInfrastructurePostgreSQLAssemblyMarker
+public interface IInfrastructurePostgreSqlAssemblyMarker
 {
 }
