@@ -4,8 +4,8 @@ using System.Threading.Tasks;
 using MediatR;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.ChangeTracking;
-using Wally.Lib.DDD.Abstractions.Commands;
 using Wally.RomMaster.FileService.Domain.Abstractions;
+using Wally.Lib.DDD.Abstractions.Commands;
 
 namespace Wally.RomMaster.FileService.Infrastructure.PipelineBehaviours;
 
@@ -38,7 +38,7 @@ public class UpdateMetadataHandlerBehavior<TRequest, TResponse> : IPipelineBehav
 		return response;
 	}
 
-	private void UpdateAggregateMetadata(IEnumerable<EntityEntry<IAggregateRoot>> entries)
+	private void UpdateAggregateMetadata(IEnumerable<EntityEntry> entries)
 	{
 		var now = _dateTimeProvider.GetDateTime();
 
@@ -50,18 +50,28 @@ public class UpdateMetadataHandlerBehavior<TRequest, TResponse> : IPipelineBehav
 					entry.CurrentValues.SetValues(
 						new Dictionary<string, object>
 						{
-							{ nameof(IAggregateRoot.ModifiedAt), now },
-							{ nameof(IAggregateRoot.ModifiedById), _userProvider.GetCurrentUserId() },
+							{
+								nameof(IAggregateRoot.ModifiedAt), now
+							},
+							{
+								nameof(IAggregateRoot.ModifiedById), _userProvider.GetCurrentUserId()
+							},
 						});
 					break;
 				case EntityState.Added:
 					entry.CurrentValues.SetValues(
 						new Dictionary<string, object>
 						{
-							{ nameof(IAggregateRoot.CreatedAt), now },
-							{ nameof(IAggregateRoot.CreatedById), _userProvider.GetCurrentUserId() },
+							{
+								nameof(IAggregateRoot.CreatedAt), now
+							},
+							{
+								nameof(IAggregateRoot.CreatedById), _userProvider.GetCurrentUserId()
+							},
 						});
 					break;
+				default:
+					continue;
 			}
 		}
 	}
