@@ -1,8 +1,8 @@
 using System.Linq;
 using FluentAssertions;
 using FluentAssertions.Execution;
-using Wally.Lib.DDD.Abstractions.Commands;
 using Wally.RomMaster.FileService.Tests.ConventionTests.Helpers;
+using Wally.Lib.DDD.Abstractions.Commands;
 using Xunit;
 
 namespace Wally.RomMaster.FileService.Tests.ConventionTests;
